+++ js/data.js
@@ -1,0 +1,34 @@
+const DESCRIPTIONS = [
+  'Прекрасный закат на море',
+  'Горный пейзаж ранним утром',
+  'Улочки старого города',
+  'Кофе в уютной кофейне',
+  'Прогулка по осеннему парку',
+  'Архитектура современного мегаполиса',
+  'Момент из путешествия',
+  'Тихий вечер дома',
+  'Яркие краски лета',
+  'Зимняя сказка в лесу',
+  'Уличное искусство и граффити',
+  'Морской бриз и пляж',
+  'Городские огни ночью',
+  'Природа в её лучшем проявлении',
+  'Моменты счастливой жизни'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = [
+  'Артём', 'Мария', 'Дмитрий', 'Анна', 'Сергей',
+  'Елена', 'Алексей', 'Ольга', 'Иван', 'Наталья',
+  'Павел', 'Виктория', 'Михаил', 'Юлия', 'Андрей'
+];
+
+export {DESCRIPTIONS, MESSAGES, NAMES};
